Refactor: rename to custom action and move them to subpackage
