+++ app/src/main/java/io/github/antoinepirlot/satunes/ui/components/buttons/playback/custom_actions/AddToPlaylistCustomAction.kt
@@ -23,7 +23,7 @@
  *  PS: I don't answer quickly.
  */
 
-package io.github.antoinepirlot.satunes.ui.components.buttons.playback
+package io.github.antoinepirlot.satunes.ui.components.buttons.playback.custom_actions
 
 import android.content.Context
 import androidx.compose.runtime.Composable
@@ -44,6 +44,7 @@ import io.github.antoinepirlot.satunes.database.services.DatabaseManager
 import io.github.antoinepirlot.satunes.icons.SatunesIcons
 import io.github.antoinepirlot.satunes.playback.services.PlaybackController
 import io.github.antoinepirlot.satunes.services.MediaSelectionManager
+import io.github.antoinepirlot.satunes.ui.components.buttons.playback.RowButton
 import io.github.antoinepirlot.satunes.ui.components.dialog.MediaSelectionDialog
 import java.util.SortedMap
 
@@ -52,7 +53,7 @@ import java.util.SortedMap
  */
 
 @Composable
-internal fun AddToPlaylistRowButton(
+internal fun AddToPlaylistCustomAction(
     modifier: Modifier = Modifier,
 ) {
     val context: Context = LocalContext.current
@@ -104,5 +105,5 @@ private fun addMusicPlayingToPlaylist(
 @Preview
 @Composable
 private fun AddToPlaylistRowButtonPreview() {
-    AddToPlaylistRowButton()
+    AddToPlaylistCustomAction()
 }
