+++ app/src/main/java/io/github/antoinepirlot/satunes/ui/views/playback/common/MusicPlayingControlView.kt
@@ -47,7 +47,7 @@ import io.github.antoinepirlot.satunes.database.models.Artist
 import io.github.antoinepirlot.satunes.playback.services.PlaybackController
 import io.github.antoinepirlot.satunes.ui.ScreenSizes
 import io.github.antoinepirlot.satunes.ui.components.bars.MusicControlBar
-import io.github.antoinepirlot.satunes.ui.components.buttons.playback.PlaybackButtonsRow
+import io.github.antoinepirlot.satunes.ui.components.buttons.playback.custom_actions.PlaybackCustomActionsRow
 import io.github.antoinepirlot.satunes.ui.components.images.MusicPlayingAlbumArtwork
 import io.github.antoinepirlot.satunes.ui.components.texts.NormalText
 import io.github.antoinepirlot.satunes.ui.components.texts.Subtitle
@@ -97,7 +97,7 @@ internal fun MusicPlayingControlView(
                 modifier = Modifier.clickable { onArtistClick(musicPlaying.value!!.artist) },
                 text = musicPlaying.value!!.artist.title
             )
-            PlaybackButtonsRow()
+            PlaybackCustomActionsRow()
             MusicControlBar()
         }
     }
